requirements.txt
--find-links https://download.pytorch.org/whl/cu118
torch>=2.0.0+cu118
torchvision>=0.15.0+cu118
torchaudio>=2.0.0+cu118
torch-geometric>=2.3.0
transformers>=4.15.0
numpy>=1.19.2
pandas>=1.2.0
networkx>=2.5
matplotlib>=3.3.4
seaborn>=0.11.2
scikit-learn>=0.24.2
pytest>=6.2.5
pyyaml>=5.4.1
tqdm>=4.62.3
spacy>=3.1.0
sympy>=1.8
beautifulsoup4>=4.9.3
requests>=2.26.0
jupyter>=1.0.0
wandb>=0.12.0
tensorboard>=2.7.0
pytorch-lightning>=1.5.0